Build 84: Code refractoring, middle value calculation.

+++ src/StartupUtils.cpp
@@ -15,7 +15,7 @@
 using namespace std;
 using namespace FilesystemProvider;
 
-int StartupUtils::grabFromCLI(double& startRef, double& endRef, double& stepRef,
+int StartupUtils::grabFromCLI(long double& startRef, long double& endRef, long double& stepRef,
 		double& pStepRef, Matrix& modelRef, int& blockCountRef, bool& randRef,
 		string& wDirRef) {
 	cout << "Do you want to init randomizer? (yes/no) ";
@@ -97,8 +97,8 @@ int StartupUtils::grabFromCLI(double& startRef, double& endRef, double& stepRef,
 	return ocode;
 }
 
-int StartupUtils::grabFromFile(double& startRef, double& endRef,
-		double& stepRef, double& pStepRef, Matrix& modelRef, int& blockCountRef,
+int StartupUtils::grabFromFile(long double& startRef, long double& endRef,
+		long double& stepRef, double& pStepRef, Matrix& modelRef, int& blockCountRef,
 		bool& randRef, string& wDirRef, bool& useCLI, string confLocation) {
 	ifstream ifs;
 	ifs.open(confLocation);
